// src/app/core/models/neo.model.ts
export interface Asteroid {
  id: string;
  name: string;
  absoluteMagnitude: number; // brillo → útil para D3
  diameterKm: number;        // radio medio → sphere en Three.js
  isPotentiallyHazardous: boolean;

  closeApproach: {
    date: string;
    velocityKps: number;     // velocidad en km/s → animaciones 3D
    missDistanceKm: number;  // cercanía → D3 impact maps
    orbitingBody: string;
  }[];

  orbit?: {
    semiMajorAxis: number;   // eje mayor → tamaño órbita
    eccentricity: number;    // excentricidad → forma órbita
    inclination: number;     // inclinación → ángulo órbita
    periodDays: number;      // periodo → animaciones orbitales
  };
}